MAINT move util funcs on tokens and ctrees to educe.ptb
